refactor: LoginPage 단위 테스트를 위한 InputField 분리
